+++ src/JsonBuffer.cpp
@@ -34,6 +34,11 @@ JsonObject &JsonBuffer::parseObject(char *json, uint8_t nestingLimit) {
   return parser.parseObject();
 }
 
+JsonVariant JsonBuffer::parse(char *json, uint8_t nestingLimit) {
+  JsonParser parser(this, json, nestingLimit);
+  return parser.parseVariant();
+}
+
 char *JsonBuffer::strdup(const char *source, size_t length) {
   size_t size = length + 1;
   char *dest = static_cast<char *>(alloc(size));